MVP for kaocha/cljs2
A new kaocha test suite type based on Chui. At the stage where it can connect
with clients, build up a test plan, and run tests. Messages from cljs.test are
propagated back and handled by the kaocha reporter.

Still needs lots of testing and fine tuning, and work to make sure all Kaocha
features are supported, but this is a good milestone.