Ensure string in `re-find` call in case of nil
Under some conditions, Safari won't report stack data in the way
`lambdaisland.chui.runner/file-and-line` expects. It blows up when
trying to detect the location of a test failure on the `re-find` calls
since at that point it doesn't have a string. This can turn a passing
test into a test that's erroring out.

This commit adds a guard to ensure `re-find` is always handed a string.
It might be `""` in the case above in Safari, probably leading to reporting a
failing line incorrectly. But the comments above suggest this hasn't
been working entirely correctly anyway. It's a no-op in Firefox and
Chrome.